Plot the likelihood of the observed age at length data

Plot the likelihood of the observed age at length data

Usage

plotAgeLikelihood(pars, age_at_length)

Arguments

pars

A list of parameters

age_at_length

Data frame of raw age-at-length observations; will be preprocessed internally by preprocess_length_at_age().

Value

A ggplot2 object suitable for display in Shiny or saving.

Examples

# In practice provide a real `age_at_length` table for the species
# p <- plotAge(pars, age_at_length = df)